+++ server/src/main/java/com/example/tyfserver/auth/config/AuthenticationArgumentResolver.java
@@ -1,8 +1,8 @@
-package com.example.tyfserver.common.domain;
+package com.example.tyfserver.auth.config;
 
-import com.example.tyfserver.common.dto.LoginMember;
-import com.example.tyfserver.common.service.AuthenticationService;
-import javax.servlet.http.HttpServletRequest;
+import com.example.tyfserver.auth.dto.LoginMember;
+import com.example.tyfserver.auth.service.AuthenticationService;
+import com.example.tyfserver.auth.util.AuthorizationExtractor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
@@ -10,6 +10,8 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+
+import javax.servlet.http.HttpServletRequest;
 
 @Component
 @RequiredArgsConstructor
@@ -24,7 +26,7 @@ public class AuthenticationArgumentResolver implements HandlerMethodArgumentReso
 
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-        NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+                                  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
         String token = AuthorizationExtractor.extract(request);
         return authenticationService.createLoginMemberByToken(token);
