+++ server/src/main/java/com/example/tyfserver/auth/config/AuthenticationInterceptor.java
@@ -1,11 +1,13 @@
-package com.example.tyfserver.common.domain;
+package com.example.tyfserver.auth.config;
 
-import com.example.tyfserver.common.service.AuthenticationService;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import com.example.tyfserver.auth.service.AuthenticationService;
+import com.example.tyfserver.auth.util.AuthorizationExtractor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 @Component
 @RequiredArgsConstructor
@@ -15,7 +17,7 @@ public class AuthenticationInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response,
-        Object handler) throws Exception {
+                             Object handler) throws Exception {
 
         String token = AuthorizationExtractor.extract(request);
         authenticationService.validateToken(token);
