+++ server/src/main/java/com/example/tyfserver/auth/service/AuthenticationService.java
@@ -1,8 +1,8 @@
-package com.example.tyfserver.common.service;
+package com.example.tyfserver.auth.service;
 
-import com.example.tyfserver.common.domain.JwtTokenProvider;
-import com.example.tyfserver.common.dto.LoginMember;
-import com.example.tyfserver.common.dto.TokenResponse;
+import com.example.tyfserver.auth.dto.LoginMember;
+import com.example.tyfserver.auth.dto.TokenResponse;
+import com.example.tyfserver.auth.util.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
