+++ server/src/test/java/com/example/tyfserver/common/domain/JwtTokenProviderTest.java
@@ -3,6 +3,7 @@ package com.example.tyfserver.common.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.example.tyfserver.auth.util.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
